David - Move tasks that were really controllers into controller package
